friendship/requirements.txt
flask_restful
requests
flask_session
google.cloud
google-auth
google-cloud-firestore
google-cloud-storage
google-cloud-pubsub
Flask
uWSGI
uwsgi